Config capteur

+++ src/ATTinyDriver/display.c
@@ -1,7 +1,7 @@
 #include "display.h"
 
 
-void NumberToDisplay(int numberToDisplay) {
+void NumberToDisplay(uint8_t numberToDisplay) {
   /*
     *  numberToDiplsay est un entier entre 0 et 9
     *  Le décodeur est relié aux 4 pins dont l'état est géré par les bits 0 à 3 de PORTA
@@ -14,7 +14,7 @@ void DisplayToUse(uint8_t display) {
   PORTB = (PORTB & 0xF8) | (1 << display);
 }
 
-void DisplayOff(){
+void DisplayOff() {
   /*
     *  Permet d'éteindre tout les afficheur en cas de mauvaise synchronisation
   */
